refactor: rename Span -> Card
Hindsight it a wonderful thing... I realise that my definition of the
Span of a vector space V was incorrect and was actually known as
cardinality. It would have been great to get this right the first time,
and I could have just left it like, but fuck it, doesn't look like
anyone but me uses this crate anyway!