feat(core): add new trait Enclose
This trait has one required method, enclose(), which takes `self` and a
reference to some other type `S` implementing `Space`, and returns a new
instance of `Self` which encloses both spaces. This was added to support
the Mixture model in the `rstat` crate which needs to generically build
a support instance from multiple types with potentially different
bounds etc.